server/requirements.txt
celery==5.2.7
django==4.1.2
djlint==1.19.1
djangorestframework==3.14.0
django-cors-headers==3.13.0
gunicorn==20.1.0
redis==4.3.4
uvicorn==0.19.0